+++ src/prompt/PromptState.cpp
@@ -103,7 +103,7 @@ void PromptState::addCompletion(const std::u16string_view item) {
 }
 
 void PromptState::addHistory(const std::u16string_view command) {
-    if (m_command_history.size() > m_max_history->m_value - 1) {
+    if (m_command_history.size() > static_cast<std::size_t>(m_max_history->m_value - 1)) {
         // The history list is full, pop the front items.
         m_command_history.pop_front();
     }
@@ -158,7 +158,7 @@ void PromptState::registerMaxHistoryCVar() {
         const auto new_size = std::clamp(m_max_history->m_value, 8, 255);
         m_max_history->m_value = new_size;
 
-        while (m_command_history.size() > new_size) {
+        while (m_command_history.size() > static_cast<std::size_t>(new_size)) {
             m_command_history.pop_front();
         }
     });
